VZ-5822 VMO metrics exporter and unit tests (#221)
* Started metrics endpoint in main.go

* Changed endpoint implementation to use http handle

* Switched to http endpoint due to istio

* Code cleanup

* Created metrics pkg

* Created metricsExporter pkg and enabled logs

* Updated third party licenses

* changed pkg name to metricsexporter

* updated third party licenses

* Changed go.mod as per previous PR

* updated licenses to be up to date with master

* updated go.sum and go.mod

* updated licenses to match go.sum

* Added exporter test and created error handling logic

* added deployment/update metrics

* added descriptions for metrics

* code cleanup

* Tested to be working

* added template functions, ready to test deployment metrics

* registered all available metrics

* minor fix in event increment template

* fix template

* remove count template

* Unit testing in-progress

* added incomplete unit tests

* added exporter file

* fixed import issue

* fixed import issue

* Fixed go mod issue

* first working version before refactor

* added member functions and structs

* completed metricsExporter.go, ready for controller integration

* all tests passing, metrics v2

* Completed changes, awaiting third party license approval

* added comments, finished unit testing, finished exporter, most metrics added

* updated third party licenses

* fixed go lint issue

* fixed second go lint issue

* Fixed 3rd go lint issue

* Updated go mod to have direct dependencies in one require block

* Fixed typo in metrics exporter test

* refactored metricsexporter

* added metricsexporter package

* Added initialization

* Completed refactor, pretesting

* All unit tests passing

* added nolint comments to make sure metric types remain private

* fixed comments for nolint

* fixed comments for nolint

* changed allmetrics array test

* added all metrics

* performed code cleanup

* version to test e2e tests

* version to test e2e tests fixed copyright

* version to test e2e tests fixed unit tests to initialize metrics

* added error handling

* Added comments and made metric types public

* Changed counter and gauge metric type name and fixed comments

* Fixed minor build issue due to function name

* Reverted change of function name ot ensure consistency between all 3 exporters

* Fixed accidental commit

* Refactored unit tests to be loop based

* Deleted newline at the end of ConfigMap.go

* Added all possible unit tests

* Added error checking for function metrics as well

* Fixed small build issue related to error handling

+++ pkg/vmo/deployment.go
@@ -12,6 +12,7 @@ import (
 	vmcontrollerv1 "github.com/verrazzano/verrazzano-monitoring-operator/pkg/apis/vmcontroller/v1"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/config"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/constants"
+	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/resources/deployments"
 	appsv1 "k8s.io/api/apps/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -41,6 +42,11 @@ func updateOpenSearchDashboardsDeployment(osd *appsv1.Deployment, controller *Co
 		err = updateDeployment(controller, vmo, existingDeployment, osd)
 	}
 	if err != nil {
+		metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+		if metricErr != nil {
+			return metricErr
+		}
+		metric.Inc()
 		controller.log.Errorf("Failed to update deployment %s/%s: %v", osd.Namespace, osd.Name, err)
 		return err
 	}
@@ -50,6 +56,15 @@ func updateOpenSearchDashboardsDeployment(osd *appsv1.Deployment, controller *Co
 
 // CreateDeployments create/update VMO deployment k8s resources
 func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, pvcToAdMap map[string]string, existingCluster bool) (dirty bool, err error) {
+	// The error count is incremented by the function which calls createDeployment
+	functionMetric, functionError := metricsexporter.GetFunctionMetrics(metricsexporter.NamesDeployment)
+	if functionError == nil {
+		functionMetric.LogStart()
+		defer functionMetric.LogEnd(false)
+	} else {
+		return false, functionError
+	}
+
 	// Assigning the following spec members seems like a hack; is any
 	// better way to make these values available where the deployments are created?
 	vmo.Spec.NatGatewayIPs = controller.operatorConfig.NatGatewayIPs
@@ -91,6 +106,11 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 			}
 		}
 		if err != nil {
+			metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+			if metricErr != nil {
+				return false, metricErr
+			}
+			metric.Inc()
 			controller.log.Errorf("Failed to update deployment %s/%s: %v", curDeployment.Namespace, curDeployment.Name, err)
 			return false, err
 		}
@@ -138,15 +158,30 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 }
 
 func deleteDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, deployment *appsv1.Deployment) error {
+	metric, metricErr := metricsexporter.GetCounterMetrics(metricsexporter.NamesDeploymentDeleteCounter)
+	if metricErr != nil {
+		return metricErr
+	}
+	metric.Inc()
 	controller.log.Debugf("Deleting deployment %s", deployment.Name)
 	err := controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Delete(context.TODO(), deployment.Name, metav1.DeleteOptions{})
 	if err != nil {
+		metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentDeleteError)
+		if metricErr != nil {
+			return metricErr
+		}
+		metric.Inc()
 		controller.log.Errorf("Failed to delete deployment %s: %v", deployment.Name, err)
 	}
 	return nil
 }
 
 func updateDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, existingDeployment, curDeployment *appsv1.Deployment) error {
+	metric, metricErr := metricsexporter.GetCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+	if metricErr != nil {
+		return metricErr
+	}
+	metric.Inc()
 	var err error
 	curDeployment.Spec.Selector = existingDeployment.Spec.Selector
 	specDiffs := diff.Diff(existingDeployment, curDeployment)
@@ -173,7 +208,11 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 		if !isUpdateAllowed(controller, vmo, current) {
 			continue
 		}
-
+		metric, metricErr := metricsexporter.GetCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+		if metricErr != nil {
+			return false, metricErr
+		}
+		metric.Inc()
 		// Selector may not change, so we copy over from existing
 		current.Spec.Selector = existing.Spec.Selector
 		// Deployment spec differences, so call Update() and return
@@ -183,6 +222,11 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 			controller.log.Oncef("Updating deployment %s in namespace %s", current.Name, current.Namespace)
 			_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), current, metav1.UpdateOptions{})
 			if err != nil {
+				metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+				if metricErr != nil {
+					return false, metricErr
+				}
+				metric.Inc()
 				return false, err
 			}
 			//okay to return dirty=false after updating the *last* deployment
@@ -212,9 +256,18 @@ func updateAllDeployments(controller *Controller, vmo *vmcontrollerv1.Verrazzano
 		if err != nil {
 			return false, err
 		}
-
+		metric, metricErr := metricsexporter.GetCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+		if metricErr != nil {
+			return false, metricErr
+		}
+		metric.Inc()
 		_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), curDeployment, metav1.UpdateOptions{})
 		if err != nil {
+			metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+			if metricErr != nil {
+				return false, metricErr
+			}
+			metric.Inc()
 			return false, err
 		}
 	}
